now able to define input file using Path, restructuring of blocks
