+++ src/opi/input/blocks/block_autoci.py
@@ -1,6 +1,6 @@
 from typing import Literal
 
-from opi.input.blocks.base import Block
+from opi.input.blocks import Block
 
 _all__ = ("BlockAutoCI",)
