+++ src/opi/input/blocks/block_cosmors.py
@@ -1,6 +1,6 @@
 from pydantic import field_validator
 
-from opi.input.blocks.base import Block, InputFilePath
+from opi.input.blocks import Block, InputFilePath
 from opi.input.simple_keywords import SimpleKeyword, Solvent
 
 __all__ = ("BlockCosmors",)
