+++ src/opi/input/blocks/block_docker.py
@@ -2,7 +2,7 @@ from typing import Literal
 
 from pydantic import field_validator
 
-from opi.input.blocks.base import Block, InputFilePath
+from opi.input.blocks import Block, InputFilePath
 
 __all__ = ("BlockDocker",)
