+++ src/opi/input/blocks/block_eda.py
@@ -1,6 +1,6 @@
 from pydantic import field_validator
 
-from opi.input.blocks.base import Block, InputFilePath
+from opi.input.blocks import Block, InputFilePath
 
 __all__ = ("BlockEda",)
