+++ src/opi/input/blocks/block_freq.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from pydantic import FilePath, field_validator
 
-from opi.input.blocks.base import Block, IntGroup
+from opi.input.blocks import Block, IntGroup
 
 __all__ = ("HessList", "BlockFreq")
