+++ src/opi/input/blocks/block_ice.py
@@ -1,6 +1,6 @@
 from typing import Literal
 
-from opi.input.blocks.base import Block
+from opi.input.blocks import Block
 
 __all__ = ("BlockIce",)
