+++ src/opi/input/blocks/block_rel.py
@@ -2,7 +2,7 @@ from typing import Literal
 
 from pydantic import field_validator
 
-from opi.input.blocks.base import Block, NumList
+from opi.input.blocks import Block, NumList
 
 __all__ = ("BlockRel",)
